feat: host english version on the gov site
Use the i18n package to translate static components such as the Header
and Footer. The English version is indicated by 'en' in the URL, while
the Ukrainian version does not appear in the URL. The content version
is determined by the language in the URL: if the URL lacks a language,
the Ukrainian version is retrieved; if it includes the 'en' prefix,
the English version is retrieved.